Port to x64 with conversion warning cleaning

+++ mimetic/contenttype.cxx
@@ -32,7 +32,7 @@ ContentType::Boundary::Boundary()
                 "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
                 "-_."; // "=+,()/"
         stringstream ss;
-        srand(time(0));
+        srand(static_cast<int>(time(0)));
         short tbSize = sizeof(tb)-1;
         for(uint i=0; i < 48; ++i)
         {
